refactor(serve): remove the PTY WebSocket relay; it's dead after the live unification
With the web dashboard rendering the capture-snapshot live view everywhere
(xterm/PTY frontend gone), the server's PTY relay is unreachable. Delete it.

- src/server/ws.rs deleted: the handle_terminal_ws relay, the /ws,
  /terminal/ws, /container-terminal/ws routes and their handlers,
  attach_command_args, resize_pty, the SessionPrimaries primary-client map
  and claim/release helpers, the SIGSTOP pause refcount, the timing
  ping/pong, and the PTY-side size-owner lock integration (pty_claim_size,
  the heartbeat task, etc.).
- The shared pane helpers that the live handlers still use moved verbatim to
  src/server/pane.rs: respawn_paired_if_dead / respawn_container_if_dead
  (dead-shell rescue), wait_for_tmux_ready / PaneReadiness / parse_pane_dead
  (the readiness probe), close_early, and the close-code constants. Their
  unit tests came along.
- AppState loses session_primaries and session_pause_counts (PTY-only).
- live_ws imports from super::pane.

The size-owner lock still exists in tmux user options and is honored by the
remaining live handlers and the native TUI; only the PTY half of it (which
nothing reached) is gone.

Verified: cargo build/clippy/fmt --features serve, full lib tests (one
pre-existing root-sandbox storage failure), live golden-path/size-owner/
no-flutter green against the rebuilt binary.
